console: Move gateway logics to new file move logics to logics/gateways.js
