+++ pkg/webui/console/store/middleware/logics/index.js
@@ -18,7 +18,6 @@ import applications from './applications'
 import devices from './devices'
 import device from './device'
 import gateways from './gateways'
-import gateway from './gateway'
 import configuration from './configuration'
 
 export default [
@@ -28,6 +27,5 @@ export default [
   ...devices,
   ...device,
   ...gateways,
-  ...gateway,
   ...configuration,
 ]
